+++ library/core/src/main/java/com/google/android/exoplayer2/extractor/DefaultExtractorsFactory.java
@@ -16,6 +16,7 @@
 package com.google.android.exoplayer2.extractor;
 
 import com.google.android.exoplayer2.extractor.amr.AmrExtractor;
+import com.google.android.exoplayer2.extractor.flac.FlacExtractor;
 import com.google.android.exoplayer2.extractor.flv.FlvExtractor;
 import com.google.android.exoplayer2.extractor.mkv.MatroskaExtractor;
 import com.google.android.exoplayer2.extractor.mp3.Mp3Extractor;
@@ -55,12 +56,13 @@ import java.lang.reflect.Constructor;
  */
 public final class DefaultExtractorsFactory implements ExtractorsFactory {
 
-  private static final Constructor<? extends Extractor> FLAC_EXTRACTOR_CONSTRUCTOR;
+  private static final Constructor<? extends Extractor> FLAC_EXTENSION_EXTRACTOR_CONSTRUCTOR;
+
   static {
-    Constructor<? extends Extractor> flacExtractorConstructor = null;
+    Constructor<? extends Extractor> flacExtensionExtractorConstructor = null;
     try {
       // LINT.IfChange
-      flacExtractorConstructor =
+      flacExtensionExtractorConstructor =
           Class.forName("com.google.android.exoplayer2.ext.flac.FlacExtractor")
               .asSubclass(Extractor.class)
               .getConstructor();
@@ -71,7 +73,7 @@ public final class DefaultExtractorsFactory implements ExtractorsFactory {
       // The FLAC extension is present, but instantiation failed.
       throw new RuntimeException("Error instantiating FLAC extension", e);
     }
-    FLAC_EXTRACTOR_CONSTRUCTOR = flacExtractorConstructor;
+    FLAC_EXTENSION_EXTRACTOR_CONSTRUCTOR = flacExtensionExtractorConstructor;
   }
 
   private boolean constantBitrateSeekingEnabled;
@@ -208,7 +210,7 @@ public final class DefaultExtractorsFactory implements ExtractorsFactory {
 
   @Override
   public synchronized Extractor[] createExtractors() {
-    Extractor[] extractors = new Extractor[FLAC_EXTRACTOR_CONSTRUCTOR == null ? 13 : 14];
+    Extractor[] extractors = new Extractor[14];
     extractors[0] = new MatroskaExtractor(matroskaFlags);
     extractors[1] = new FragmentedMp4Extractor(fragmentedMp4Flags);
     extractors[2] = new Mp4Extractor(mp4Flags);
@@ -237,13 +239,16 @@ public final class DefaultExtractorsFactory implements ExtractorsFactory {
                     ? AmrExtractor.FLAG_ENABLE_CONSTANT_BITRATE_SEEKING
                     : 0));
     extractors[12] = new Ac4Extractor();
-    if (FLAC_EXTRACTOR_CONSTRUCTOR != null) {
+    // Prefer the FLAC extension extractor because it supports seeking.
+    if (FLAC_EXTENSION_EXTRACTOR_CONSTRUCTOR != null) {
       try {
-        extractors[13] = FLAC_EXTRACTOR_CONSTRUCTOR.newInstance();
+        extractors[13] = FLAC_EXTENSION_EXTRACTOR_CONSTRUCTOR.newInstance();
       } catch (Exception e) {
         // Should never happen.
         throw new IllegalStateException("Unexpected error creating FLAC extractor", e);
       }
+    } else {
+      extractors[13] = new FlacExtractor();
     }
     return extractors;
   }